requirements.txt
clang-5==5.0.post2
decorator==4.4.0
networkx==2.3
